feat(storage-v2): update posix reader config to follow same structre as posix writer, but only with path as only required config for reader

+++ sda/internal/storage/v2/posix/reader/reader_test.go
@@ -124,6 +124,6 @@ func (ts *ReaderTestSuite) TestGetFileSize() {
 		ts.FailNow(err.Error())
 	}
 
-	ts.Equal(len("file 6 content in dir2"), fileSize)
+	ts.Equal(int64(len("file 6 content in dir2")), fileSize)
 
 }
